orchestrator/requirements.txt
quart==0.19.9
quart-cors==0.7.0
aiohttp==3.11.10
psycopg2-binary==2.9.9
pika==1.3.1
prometheus-client==0.19.0
python-jose==3.3.0
requests==2.31.0
websockets==12.0
python-dotenv==1.0.0
PyJWT==2.3.0
simple-websocket==0.10.1
Werkzeug>=3.0.0